:recycle: refactor: Refatorar a estrutura atual da parte instituições e aplicando componentização
