+++ civitas-frontend/src/app/dashboard/instituicoes/_components/index.ts
@@ -1,0 +1,4 @@
+export { default as InstituicaoInsightsModal } from "./InstituicaoInsightsModal";
+export { default as InstituicoesErrorAlert } from "./InstituicoesErrorAlert";
+export { default as InstituicoesOverviewSection } from "./InstituicoesOverviewSection";
+export { default as InstituicoesTableSection } from "./InstituicoesTableSection";
