+++ civitas-frontend/src/app/dashboard/instituicoes/_utils/formatters.ts
@@ -1,0 +1,6 @@
+export const formatCurrency = (value: number): string => {
+  return new Intl.NumberFormat("pt-BR", {
+    style: "currency",
+    currency: "BRL",
+  }).format(value);
+};
